move more common logic to parent class
